squash: refactor portal organization
