+++ protocol-designer/src/components/StepEditForm/ConfirmDeleteModal.js
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react'
 import {ContinueModal} from '@opentrons/components'
-import ConfirmModalPortal from '../modals/ConfirmModalPortal'
+import {Portal} from '../portals/MainPageModalPortal'
 import modalStyles from '../modals/modal.css'
 
 const CONFIRM_DELETE_TEXT = 'Are you sure you want to delete this step?'
@@ -10,11 +10,11 @@ type Props = $Diff<React.ElementProps<typeof ContinueModal>, {children: *}>
 
 function ConfirmDeleteModal (props: Props) {
   return (
-    <ConfirmModalPortal>
+    <Portal>
       <ContinueModal className={modalStyles.modal} {...props}>
         {CONFIRM_DELETE_TEXT}
       </ContinueModal>
-    </ConfirmModalPortal>
+    </Portal>
   )
 }
