+++ protocol-designer/src/constants.js
@@ -46,5 +46,3 @@ export function getAllWellsForLabware (labwareType: string): Array<string> {
 }
 
 export const FIXED_TRASH_ID: 'trashId' = 'trashId'
-
-export const CONFIRM_MODAL_ROOT_ID = 'confirm-modal-root'
